Add windows support (#2)
It now works for both linux and windows, but the execution differs:
For both platforms, a virtual python environment is created and the packages are installed in there.

On linux, Gimp`s native python interpreter is used to execute sam by adding the additional packages to the path.

On windows, this does not work, therefore the gimp script exports the image as png, and the python of the virtual environment is used to perform the inference and saves it to another png, which is then imported back into gimp,